--- src/main/java/kr/hhplus/be/server/controller/response/CurrentPointResponse.java
@@ -1,14 +1,0 @@
-package kr.hhplus.be.server.controller.response;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(description = "포인트 조회 응답")
-public record CurrentPointResponse(
-    @Schema(description = "유저 ID")
-    Long userId,
-
-    @Schema(description = "포인트 잔액")
-    Long amount
-) {
-
-}
